# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: Nodeo API
  description: Private self-hosted media streaming service
  version: 2.0.0

servers:
  - url: /api

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    LoginResponse:
      type: object
      required: [token, expiresIn]
      properties:
        token:
          type: string
        expiresIn:
          type: string

    UserInfo:
      type: object
      required: [username]
      properties:
        username:
          type: string

    HealthResponse:
      type: object
      required: [status, timestamp]
      properties:
        status:
          type: string
        timestamp:
          type: string
          format: date-time

    Collection:
      type: object
      required: [id, name, is_encrypted, is_unlocked, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        is_encrypted:
          type: boolean
        is_unlocked:
          type: boolean
          description: Whether the collection is currently unlocked (always true for non-encrypted)
        cover_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CollectionCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
        passphrase:
          type: string
          description: If provided, the collection will be encrypted with this passphrase

    CollectionUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string

    UnlockRequest:
      type: object
      required: [passphrase]
      properties:
        passphrase:
          type: string

    UnlockResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    MediaMetadata:
      type: object
      description: Normalized tags extracted from the uploaded file (ffprobe), before transcoding
      properties:
        title:
          type: string
          nullable: true
        artist:
          type: string
          nullable: true
        album:
          type: string
          nullable: true
        album_artist:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        year:
          type: string
          nullable: true
        track:
          type: string
          nullable: true
        composer:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        encoder:
          type: string
          nullable: true
        creation_time:
          type: string
          nullable: true

    MediaProcessingProgress:
      type: object
      required: [stage, overall_percent]
      description: Live job progress while status is processing (in-memory on the backend)
      properties:
        stage:
          type: string
          enum: [transcoding, thumbnails, encrypting, uploading_main, uploading_thumbs]
        overall_percent:
          type: integer
          minimum: 0
          maximum: 100
        current_sec:
          type: integer
          nullable: true
          description: Encoder timeline position during transcoding when known
        total_sec:
          type: integer
          nullable: true
          description: Estimated source duration used for transcoding progress

    Media:
      type: object
      required: [id, collection_id, title, media_type, profile, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        collection_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        media_type:
          type: string
          enum: [video, audio]
        duration_sec:
          type: integer
          nullable: true
        file_size_bytes:
          type: integer
          format: int64
          nullable: true
        profile:
          type: string
        mime_type:
          type: string
          nullable: true
        status:
          type: string
          enum: [processing, ready, error]
        created_at:
          type: string
          format: date-time
        thumb_frame_count:
          type: integer
          description: Number of video preview frames (0 for audio or processing)
          minimum: 0
        metadata:
          allOf:
            - $ref: '#/components/schemas/MediaMetadata'
          nullable: true
          description: ID3 / container tags from the original upload when ffprobe could read them
        progress:
          allOf:
            - $ref: '#/components/schemas/MediaProcessingProgress'
          nullable: true
          description: Only while status is processing; null if the server restarted mid-job

    UploadResponse:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string

paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      tags: [system]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /auth/login:
    post:
      operationId: login
      summary: Authenticate user
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Missing credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/me:
    get:
      operationId: getMe
      summary: Get current user info
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /collections:
    get:
      operationId: listCollections
      summary: List all collections
      tags: [collections]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collection'
    post:
      operationId: createCollection
      summary: Create a new collection
      tags: [collections]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
      responses:
        '201':
          description: Collection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'

  /collections/{id}:
    get:
      operationId: getCollection
      summary: Get a collection by ID
      tags: [collections]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Collection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: updateCollection
      summary: Update a collection
      tags: [collections]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionUpdate'
      responses:
        '200':
          description: Collection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
    delete:
      operationId: deleteCollection
      summary: Delete a collection
      tags: [collections]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Collection deleted

  /collections/{id}/unlock:
    post:
      operationId: unlockCollection
      summary: Unlock an encrypted collection with passphrase
      tags: [collections]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnlockRequest'
      responses:
        '200':
          description: Collection unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnlockResponse'
        '403':
          description: Wrong passphrase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /collections/{id}/media:
    get:
      operationId: listCollectionMedia
      summary: List media in a collection
      tags: [media]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Media'
    post:
      operationId: uploadMedia
      summary: Upload a media file to a collection
      tags: [media]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                description:
                  type: string
                profile:
                  type: string
                  enum: ['480p', '720p', '1080p', '1080p60', 'mp3-128', 'mp3-192', 'mp3-320', 'aac-256']
                  default: '720p'
      responses:
        '202':
          description: Upload accepted, processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '403':
          description: Collection is locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /media/{id}:
    get:
      operationId: getMedia
      summary: Get media details
      description: >
        When status is `processing`, includes optional `progress` (stage, overall_percent, optional time hints).
        Responses use `Cache-Control: private, no-store` while processing so clients see fresh progress.
      tags: [media]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Media details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        '404':
          description: Media not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteMedia
      summary: Delete a media file
      tags: [media]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Media deleted

  /media/{id}/stream:
    get:
      operationId: streamMedia
      summary: Stream media content
      tags: [media]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Media stream
          content:
            video/mp4:
              schema:
                type: string
                format: binary
            audio/mpeg:
              schema:
                type: string
                format: binary
        '206':
          description: Partial content (range request)

  /media/{id}/thumb:
    get:
      operationId: getMediaThumb
      summary: Get media thumbnail frame (JPEG)
      tags: [media]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: i
          in: query
          description: Frame index from 0 (multiple frames for video hover preview)
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Thumbnail image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary